% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{Beta}
\alias{Beta}
\title{Non-central beta distribution}
\description{
A R6 class to represent a non-central beta distribution.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{a}}{Get or set the value of \code{a}.}

\item{\code{b}}{Get or set the value of \code{b}.}

\item{\code{delta}}{Get or set the value of \code{delta}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Beta-new}{\code{Beta$new()}}
\item \href{#method-Beta-d}{\code{Beta$d()}}
\item \href{#method-Beta-p}{\code{Beta$p()}}
\item \href{#method-Beta-q}{\code{Beta$q()}}
\item \href{#method-Beta-r}{\code{Beta$r()}}
\item \href{#method-Beta-mean}{\code{Beta$mean()}}
\item \href{#method-Beta-median}{\code{Beta$median()}}
\item \href{#method-Beta-mode}{\code{Beta$mode()}}
\item \href{#method-Beta-sd}{\code{Beta$sd()}}
\item \href{#method-Beta-variance}{\code{Beta$variance()}}
\item \href{#method-Beta-skewness}{\code{Beta$skewness()}}
\item \href{#method-Beta-kurtosis}{\code{Beta$kurtosis()}}
\item \href{#method-Beta-kurtosisExcess}{\code{Beta$kurtosisExcess()}}
\item \href{#method-Beta-clone}{\code{Beta$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Beta-new"></a>}}
\if{latex}{\out{\hypertarget{method-Beta-new}{}}}
\subsection{Method \code{new()}}{
New beta distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Beta$new(a, b, delta)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a, b}}{shape parameters, \code{> 0}}

\item{\code{delta}}{non-centrality parameter, \code{>= 0}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{Beta} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Beta-d"></a>}}
\if{latex}{\out{\hypertarget{method-Beta-d}{}}}
\subsection{Method \code{d()}}{
Density function of the beta distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Beta$d(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{numeric vector}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The density evaluated at \code{x}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Beta-p"></a>}}
\if{latex}{\out{\hypertarget{method-Beta-p}{}}}
\subsection{Method \code{p()}}{
Cumulative distribution function of the beta
  distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Beta$p(q, lower = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{q}}{numeric vector of quantiles}

\item{\code{lower}}{Boolean, whether to deal with the lower tail}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The cumulative probabilities corresponding to \code{q}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Beta-q"></a>}}
\if{latex}{\out{\hypertarget{method-Beta-q}{}}}
\subsection{Method \code{q()}}{
Quantile function of the beta distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Beta$q(p, lower = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{p}}{numeric vector of probabilities}

\item{\code{lower}}{Boolean, whether to deal with the lower tail}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The quantiles corresponding to \code{p}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Beta-r"></a>}}
\if{latex}{\out{\hypertarget{method-Beta-r}{}}}
\subsection{Method \code{r()}}{
Sampling from the beta distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Beta$r(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of simulations}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric vector of length \code{n}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Beta-mean"></a>}}
\if{latex}{\out{\hypertarget{method-Beta-mean}{}}}
\subsection{Method \code{mean()}}{
Mean of the beta distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Beta$mean()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The mean of the beta distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Beta-median"></a>}}
\if{latex}{\out{\hypertarget{method-Beta-median}{}}}
\subsection{Method \code{median()}}{
Median of the beta distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Beta$median()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The median of the beta distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Beta-mode"></a>}}
\if{latex}{\out{\hypertarget{method-Beta-mode}{}}}
\subsection{Method \code{mode()}}{
Mode of the beta distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Beta$mode()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The mode of the beta distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Beta-sd"></a>}}
\if{latex}{\out{\hypertarget{method-Beta-sd}{}}}
\subsection{Method \code{sd()}}{
Standard deviation of the beta distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Beta$sd()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The standard deviation of the beta distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Beta-variance"></a>}}
\if{latex}{\out{\hypertarget{method-Beta-variance}{}}}
\subsection{Method \code{variance()}}{
Variance of the beta distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Beta$variance()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The variance of the beta distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Beta-skewness"></a>}}
\if{latex}{\out{\hypertarget{method-Beta-skewness}{}}}
\subsection{Method \code{skewness()}}{
Skewness of the beta distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Beta$skewness()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The skewness of the beta distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Beta-kurtosis"></a>}}
\if{latex}{\out{\hypertarget{method-Beta-kurtosis}{}}}
\subsection{Method \code{kurtosis()}}{
Kurtosis of the beta distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Beta$kurtosis()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The kurtosis of the beta distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Beta-kurtosisExcess"></a>}}
\if{latex}{\out{\hypertarget{method-Beta-kurtosisExcess}{}}}
\subsection{Method \code{kurtosisExcess()}}{
Kurtosis excess of the beta distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Beta$kurtosisExcess()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The kurtosis excess of the beta distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Beta-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Beta-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Beta$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
