% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig.R
\name{inverseGaussian}
\alias{inverseGaussian}
\title{Inverse Gaussian distribution}
\description{
A R6 class to represent an inverse Gaussian distribution.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-inverseGaussian-new}{\code{inverseGaussian$new()}}
\item \href{#method-inverseGaussian-d}{\code{inverseGaussian$d()}}
\item \href{#method-inverseGaussian-p}{\code{inverseGaussian$p()}}
\item \href{#method-inverseGaussian-q}{\code{inverseGaussian$q()}}
\item \href{#method-inverseGaussian-r}{\code{inverseGaussian$r()}}
\item \href{#method-inverseGaussian-mean}{\code{inverseGaussian$mean()}}
\item \href{#method-inverseGaussian-median}{\code{inverseGaussian$median()}}
\item \href{#method-inverseGaussian-mode}{\code{inverseGaussian$mode()}}
\item \href{#method-inverseGaussian-variance}{\code{inverseGaussian$variance()}}
\item \href{#method-inverseGaussian-skewness}{\code{inverseGaussian$skewness()}}
\item \href{#method-inverseGaussian-kurtosis}{\code{inverseGaussian$kurtosis()}}
\item \href{#method-inverseGaussian-kurtosisExcess}{\code{inverseGaussian$kurtosisExcess()}}
\item \href{#method-inverseGaussian-clone}{\code{inverseGaussian$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inverseGaussian-new"></a>}}
\if{latex}{\out{\hypertarget{method-inverseGaussian-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inverseGaussian$new(mu, lambda)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inverseGaussian-d"></a>}}
\if{latex}{\out{\hypertarget{method-inverseGaussian-d}{}}}
\subsection{Method \code{d()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inverseGaussian$d(x, log = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inverseGaussian-p"></a>}}
\if{latex}{\out{\hypertarget{method-inverseGaussian-p}{}}}
\subsection{Method \code{p()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inverseGaussian$p(q, lower = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inverseGaussian-q"></a>}}
\if{latex}{\out{\hypertarget{method-inverseGaussian-q}{}}}
\subsection{Method \code{q()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inverseGaussian$q(p, lower = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inverseGaussian-r"></a>}}
\if{latex}{\out{\hypertarget{method-inverseGaussian-r}{}}}
\subsection{Method \code{r()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inverseGaussian$r(n)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inverseGaussian-mean"></a>}}
\if{latex}{\out{\hypertarget{method-inverseGaussian-mean}{}}}
\subsection{Method \code{mean()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inverseGaussian$mean()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inverseGaussian-median"></a>}}
\if{latex}{\out{\hypertarget{method-inverseGaussian-median}{}}}
\subsection{Method \code{median()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inverseGaussian$median()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inverseGaussian-mode"></a>}}
\if{latex}{\out{\hypertarget{method-inverseGaussian-mode}{}}}
\subsection{Method \code{mode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inverseGaussian$mode()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inverseGaussian-variance"></a>}}
\if{latex}{\out{\hypertarget{method-inverseGaussian-variance}{}}}
\subsection{Method \code{variance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inverseGaussian$variance()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inverseGaussian-skewness"></a>}}
\if{latex}{\out{\hypertarget{method-inverseGaussian-skewness}{}}}
\subsection{Method \code{skewness()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inverseGaussian$skewness()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inverseGaussian-kurtosis"></a>}}
\if{latex}{\out{\hypertarget{method-inverseGaussian-kurtosis}{}}}
\subsection{Method \code{kurtosis()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inverseGaussian$kurtosis()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inverseGaussian-kurtosisExcess"></a>}}
\if{latex}{\out{\hypertarget{method-inverseGaussian-kurtosisExcess}{}}}
\subsection{Method \code{kurtosisExcess()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inverseGaussian$kurtosisExcess()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inverseGaussian-clone"></a>}}
\if{latex}{\out{\hypertarget{method-inverseGaussian-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inverseGaussian$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
