% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig.R
\name{inverseGaussian}
\alias{inverseGaussian}
\title{Inverse Gaussian distribution}
\description{
A R6 class to represent an inverse Gaussian distribution.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{mu}}{Get or set the value of \code{mu}.}

\item{\code{lambda}}{Get or set the value of \code{lambda}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-inverseGaussian-new}{\code{inverseGaussian$new()}}
\item \href{#method-inverseGaussian-d}{\code{inverseGaussian$d()}}
\item \href{#method-inverseGaussian-p}{\code{inverseGaussian$p()}}
\item \href{#method-inverseGaussian-q}{\code{inverseGaussian$q()}}
\item \href{#method-inverseGaussian-r}{\code{inverseGaussian$r()}}
\item \href{#method-inverseGaussian-mean}{\code{inverseGaussian$mean()}}
\item \href{#method-inverseGaussian-median}{\code{inverseGaussian$median()}}
\item \href{#method-inverseGaussian-mode}{\code{inverseGaussian$mode()}}
\item \href{#method-inverseGaussian-variance}{\code{inverseGaussian$variance()}}
\item \href{#method-inverseGaussian-skewness}{\code{inverseGaussian$skewness()}}
\item \href{#method-inverseGaussian-kurtosis}{\code{inverseGaussian$kurtosis()}}
\item \href{#method-inverseGaussian-kurtosisExcess}{\code{inverseGaussian$kurtosisExcess()}}
\item \href{#method-inverseGaussian-clone}{\code{inverseGaussian$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inverseGaussian-new"></a>}}
\if{latex}{\out{\hypertarget{method-inverseGaussian-new}{}}}
\subsection{Method \code{new()}}{
New inverse Gaussian distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inverseGaussian$new(mu, lambda)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mu}}{location parameter}

\item{\code{lambda}}{scale parameter, \code{>0}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{inverseGaussian} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inverseGaussian-d"></a>}}
\if{latex}{\out{\hypertarget{method-inverseGaussian-d}{}}}
\subsection{Method \code{d()}}{
Density function of the inverse Gaussian distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inverseGaussian$d(x, log = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{numeric vector}

\item{\code{log}}{Boolean, whether to return the logarithm of the density}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The density or the log-density evaluated at \code{x}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inverseGaussian-p"></a>}}
\if{latex}{\out{\hypertarget{method-inverseGaussian-p}{}}}
\subsection{Method \code{p()}}{
Cumulative distribution function of the inverse Gaussian
  distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inverseGaussian$p(q, lower = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{q}}{numeric vector of quantiles}

\item{\code{lower}}{Boolean, whether to deal with the lower tail}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The cumulative probabilities corresponding to \code{q}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inverseGaussian-q"></a>}}
\if{latex}{\out{\hypertarget{method-inverseGaussian-q}{}}}
\subsection{Method \code{q()}}{
Quantile function of the inverse Gaussian distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inverseGaussian$q(p, lower = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{p}}{numeric vector of probabilities}

\item{\code{lower}}{Boolean, whether to deal with the lower tail}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The quantiles corresponding to \code{p}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inverseGaussian-r"></a>}}
\if{latex}{\out{\hypertarget{method-inverseGaussian-r}{}}}
\subsection{Method \code{r()}}{
Sampling from the inverse Gaussian distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inverseGaussian$r(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of simulations}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric vector of length \code{n}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inverseGaussian-mean"></a>}}
\if{latex}{\out{\hypertarget{method-inverseGaussian-mean}{}}}
\subsection{Method \code{mean()}}{
Mean of the inverse Gaussian distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inverseGaussian$mean()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The mean of the inverse Gaussian distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inverseGaussian-median"></a>}}
\if{latex}{\out{\hypertarget{method-inverseGaussian-median}{}}}
\subsection{Method \code{median()}}{
Median of the inverse Gaussian distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inverseGaussian$median()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The median of the inverse Gaussian distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inverseGaussian-mode"></a>}}
\if{latex}{\out{\hypertarget{method-inverseGaussian-mode}{}}}
\subsection{Method \code{mode()}}{
Mode of the inverse Gaussian distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inverseGaussian$mode()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The mode of the inverse Gaussian distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inverseGaussian-variance"></a>}}
\if{latex}{\out{\hypertarget{method-inverseGaussian-variance}{}}}
\subsection{Method \code{variance()}}{
Variance of the inverse Gaussian distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inverseGaussian$variance()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The variance of the inverse Gaussian distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inverseGaussian-skewness"></a>}}
\if{latex}{\out{\hypertarget{method-inverseGaussian-skewness}{}}}
\subsection{Method \code{skewness()}}{
Skewness of the inverse Gaussian distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inverseGaussian$skewness()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The skewness of the inverse Gaussian distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inverseGaussian-kurtosis"></a>}}
\if{latex}{\out{\hypertarget{method-inverseGaussian-kurtosis}{}}}
\subsection{Method \code{kurtosis()}}{
Kurtosis of the inverse Gaussian distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inverseGaussian$kurtosis()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The kurtosis of the inverse Gaussian distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inverseGaussian-kurtosisExcess"></a>}}
\if{latex}{\out{\hypertarget{method-inverseGaussian-kurtosisExcess}{}}}
\subsection{Method \code{kurtosisExcess()}}{
Kurtosis excess of the inverse Gaussian distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inverseGaussian$kurtosisExcess()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The kurtosis excess of the inverse Gaussian distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inverseGaussian-clone"></a>}}
\if{latex}{\out{\hypertarget{method-inverseGaussian-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{inverseGaussian$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
